- apply importance multiplier to zone T&R requests based on Zone Priority (None=0, Low=0.1, Normal=1, High=5)
- handle decimal requests correctly

Related work items: #18308

+++ algos/src/main/java/a75/io/algos/tr/TrimResponseProcessor.java
@@ -53,7 +53,7 @@ public class TrimResponseProcessor
     
         double sp = setPoint + trSetting.getSPtrim();
         
-        int netRequests = trSetting.getR() - trSetting.getI();
+        double netRequests = trSetting.getR() - trSetting.getI();
         
         //TR System responds only when the net request count is positive
         if (netRequests > 0) {
